// client/src/local/types/property.ts
export interface BasePropertyFilters {
  query: string;
  location: string;
  priceMin: number | null;
  priceMax: number | null;
  verified: boolean;
  category?: "residential" | "commercial" | "land" | null;
}